+++ hdfs/scripts/start-hdfs.sh
@@ -11,3 +11,4 @@ tail -F -n 1000 /opt/hadoop/logs/hadoop-root-namenode-*.out &
 tail -F -n 1000 /opt/hadoop/logs/hadoop-root-datanode-*.out &
 
 /wait-for-it.sh -h localhost -p 9820 -t $WAIT_TIMEOUT_SECONDS
+echo "--------------- HDFS NODES READY ---------------"